[fix] views

+++ views/admin_user/view_users.php
@@ -6,6 +6,7 @@
             <div class="col-sm-4">
                 <div class="left-sidebar">
                     <h2>List of users</h2>
+                    <a href="create/" class="btn btn-success">Create new user</a>
                     <table class="table table-striped">
                         <thead>
                             <tr>
@@ -19,16 +20,18 @@
                         </thead>
                         <tbody>
                             <?php foreach ($users as $userList): ?>
-                                <tr>
-                                    <td><p class="card-title"><?php echo $userList['id']; ?></p></td>
-                                    <td><p class="card-title"><?php echo $userList['name']; ?></p></td>
-                                    <td><p class="card-title"><?php echo $userList['email']; ?></p></td>
-                                    <td><p class="card-title"><?php echo $userList['phone']; ?></p></td>
-                                    <td><p class="card-title"><?php echo $userList['balance']; ?></p></td>
-                                    <td><p class="card-title"><?php echo $userList['password']; ?></p></td>
-                                    <td><a href="update/<?php echo $userList['id']; ?>" class="btn btn-primary">Edit</a></td>
-                                    <td><a href="delete/<?php echo $userList['id']; ?>" class="btn btn-danger">Delete</a></td>
-                                </tr>
+                                <?php if (User::getRole($userList['id']) != 'admin'): ?>
+                                    <tr>
+                                        <td><p class="card-title"><?php echo $userList['id']; ?></p></td>
+                                        <td><p class="card-title"><?php echo $userList['name']; ?></p></td>
+                                        <td><p class="card-title"><?php echo $userList['email']; ?></p></td>
+                                        <td><p class="card-title"><?php echo $userList['phone']; ?></p></td>
+                                        <td><p class="card-title"><?php echo $userList['balance']; ?></p></td>
+                                        <td><p class="card-title"><?php echo $userList['password']; ?></p></td>
+                                        <td><a href="update/<?php echo $userList['id']; ?>" class="btn btn-primary">Edit</a></td>
+                                        <td><a href="delete/<?php echo $userList['id']; ?>" class="btn btn-danger">Delete</a></td>
+                                    </tr>
+                                <?php endif; ?>
                             <?php endforeach; ?>
                         </tbody>
                     </table>
